requirements.txt
aiogram>=3.0.0
aiohttp>=3.9.0
aiohttp-socks>=0.9.0
beautifulsoup4>=4.12.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
playwright>=1.40.0
playwright-stealth>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
openpyxl>=3.1.0
